renamed stubs to fixtures, removed unsafeCache: false from test helpers, make a es2015 fixture
