+++ test/fixtures/entry.js
