+++ test/fixtures/es2015/dep.js
